♻️ refactor: 안쓰는 코드 정리 및 컴포넌트 정리
